requirements.txt
certifi==2018.8.24
dj-database-url==0.5.0
Django==1.10.5
django-autoslug==1.9.3
django-heroku==0.3.1
django-recaptcha==1.2.1
django-taggit==0.21.3
django-widget-tweaks==1.4.1
djangorestframework==3.8.2
gunicorn==19.9.0
Markdown==3.0.1
psycopg2==2.7.5
pytz==2018.5
whitenoise==4.1
wincertstore==0.2
